refactor(batch-b): split watcher and chain hotspots into focused submodules
Continue applying the thin-facade pattern across Batch B. Each touched
hotspot file drops well under ~300 lines, seams follow responsibility,
and public API surface is unchanged.

dcc-mcp-skills watcher.rs (634 -> 256 lines) re-exports from:
  - watcher_inner.rs — WatcherInner shared state (snapshot, watched
    paths, debounce atomic, on-reload callbacks) and WatcherError.
  - watcher_filter.rs — should_reload / is_skill_related FS event
    heuristics.
  - watcher_python.rs — PySkillWatcher PyO3 wrapper (feature-gated on
    python-bindings).
  - watcher_tests.rs — unit tests for the three modules above.

dcc-mcp-actions chain.rs (638 -> 67 lines) re-exports from:
  - chain_types.rs — ChainStepResult, ChainResult, ErrorAction, and the
    crate-internal ChainStep / StepParams / ErrorHandlerFn types.
  - chain_interpolate.rs — {key} placeholder interpolation, context
    merge, and the public `context()` builder helper.
  - chain_exec.rs — ActionChain fluent builder + run() executor.
  - chain_tests.rs — unit tests for the three modules above.

Follow-up fixes to Batch B Wave 1 surfaced by a --features
python-bindings workspace check:
  - registry/python.rs now imports ExecutionMode / ThreadAffinity /
    ToolAnnotations / NextTools from dcc_mcp_models (previously brought
    in transitively through pymethods that no longer lived on
    registry/mod.rs).
  - registry/meta.rs drops an unused `pyo3::prelude::*` import.
  - validator_python.rs drops the unused `IssueCategory` import.

Docs:
  - Architecture guide records the new maintainer layouts for chain.rs
    (under dcc-mcp-actions) and watcher.rs (under dcc-mcp-skills) so
    agents can locate the right sibling file before editing.
